+++ index.html
@@ -9,6 +9,6 @@
 </head>
 <body>
 	<canvas id="ascii-canvas" width="800" height="600"></canvas>
-	<script type="text/javascript" src="ascii-canvas.js"></script>
+	<script type="module" src="main.js"></script>
 </body>
 </html>
