+++ lib/utils.js
@@ -1,0 +1,17 @@
+export function readTextFile(file, done)
+{
+  var rawFile = new XMLHttpRequest();
+  rawFile.open("GET", file, false);
+  rawFile.onreadystatechange = function ()
+  {
+      if(rawFile.readyState === 4)
+      {
+          if(rawFile.status === 200 || rawFile.status == 0)
+          {
+              var allText = rawFile.responseText;
+              done(allText);
+          }
+      }
+  }
+  rawFile.send(null);
+}
